src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/tab-order-iframe-and-shadow-expected.txt

CONSOLE ERROR: line 28: Uncaught TypeError: shadowHost.createShadowRoot is not a function
This tests that pressing Tab key should traverse into iframe and shadow tree, and pressing Shift-Tab should reverse the order.

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
